+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.4] - 2025-12-27
+
+### Added
+- Cancel sessions with `/stop`, `/cancel`, `stop`, or `cancel` commands in thread
+- Cancel sessions by reacting with ❌ or 🛑 to any post in the thread
+
 ## [0.3.3] - 2025-12-27
 
 ### Added